+++ CHANGELOG.md
@@ -21,7 +21,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
-- Renamed `UsbMidiEventPacket::from_midi` function to `UsbMidiEventPacket::from_message`.
 - Renamed `MidiClass::send_message` function to `MidiClass::send_packet`.
 - Renamed `midi_device` module to `class`.
 - Renamed `usb_midi` module to `packet` and moved it into crate root.
@@ -36,6 +35,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - `UsbMidiEventPacket::cable_number` field, use function instead.
 - `UsbMidiEventPacket::message` field, use function instead.
+- `UsbMidiEventPacket::from_midi` function, use `Message::into_packet` instead.
 
 ## [0.3.0] - 2024-05-27